move userStore to typescript
